--- a/composer.json
+++ b/composer.json
@@ -7,6 +7,7 @@
   "keywords": ["module", "xp"],
   "require" : {
     "xp-framework/core": "^11.0 | ^10.0 | ^9.0 | ^8.0 | ^7.0",
+    "xp-framework/reflection": "^2.0",
     "xp-forge/web": "^3.0 | ^2.0 | ^1.0",
     "xp-forge/marshalling": "^1.0 | ^0.3 | ^0.2",
     "xp-forge/json": "^5.0 | ^4.0 | ^3.1",