--- a/src/main/php/web/rest/MethodsIn.class.php
+++ b/src/main/php/web/rest/MethodsIn.class.php
@@ -1,5 +1,7 @@
 <?php namespace web\rest;
 
+use lang\Reflection;
+
 /**
  * Creates routing based on a given instance
  */
@@ -7,8 +9,12 @@ class MethodsIn extends Delegates {
 
   /** @param object $instance */
   public function __construct($instance) {
-    $class= typeof($instance);
-    $this->with($instance, $class->hasAnnotation('resource') ? $class->getAnnotation('resource') ?? '' : '/');
+    $class= Reflection::type($instance);
+    if ($annotation= $class->annotation(Resource::class)) {
+      $this->with($instance, (string)$annotation->argument(0));
+    } else {
+      $this->with($instance, '/');
+    }
     uksort($this->patterns, function($a, $b) { return strlen($b) - strlen($a); });
   }
 }