--- a/src/main/php/web/rest/ResourcesIn.class.php
+++ b/src/main/php/web/rest/ResourcesIn.class.php
@@ -1,11 +1,11 @@
 <?php namespace web\rest;
 
-use lang\reflect\Package;
+use lang\reflection\Package;
 
 /**
  * Creates routing based on resource classes in a given package
  *
- * @test  xp://web.rest.unittest.ResourcesInTest
+ * @test  web.rest.unittest.ResourcesInTest
  */
 class ResourcesIn extends Delegates {
 
@@ -16,10 +16,10 @@ class ResourcesIn extends Delegates {
    * @param  function(lang.XPClass): object $new Optional function to create instances
    */
   public function __construct($package, $new= null) {
-    $p= $package instanceof Package ? $package : Package::forName($package);
-    foreach ($p->getClasses() as $class) {
-      if ($class->hasAnnotation('resource')) {
-        $this->with($new ? $new($class) : $class->newInstance(), $class->getAnnotation('resource') ?? '');
+    $p= $package instanceof Package ? $package : new Package($package);
+    foreach ($p->types() as $type) {
+      if ($resource= $type->annotation(Resource::class)) {
+        $this->with($new ? $new($type->class()) : $type->newInstance(), (string)$resource->argument(0));
       }
     }
     uksort($this->patterns, function($a, $b) { return strlen($b) - strlen($a); });