This pull request migrates the code to use `xp-framework/reflection`. As this changes the API, we will releases this as a major release, 4.0.0 at the time of writing.

## Performance

No measurable impact.

### Before

```bash
$ xp test src/test/php
# ...
Test cases:  121 succeeded, 1 skipped
Memory used: 4291.82 kB (4345.94 kB peak)
Time taken:  0.061 seconds (0.182 seconds overall)
```

### After

```bash
$ xp test src/test/php
# ...
Test cases:  121 succeeded, 1 skipped
Memory used: 4406.82 kB (4496.78 kB peak)
Time taken:  0.064 seconds (0.188 seconds overall)
```

## See also

* https://github.com/xp-forge/frontend/pull/37
* https://github.com/xp-framework/rfc/issues/338
* https://github.com/xp-framework/core/pull/332