@thekid 2023-07-25
Released in https://github.com/xp-forge/rest-api/releases/tag/v4.0.0